+++ srcs/libwati/wati_abs.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   wati_abs.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bedarenn <bedarenn@student.42angouleme.fr  +#+  +:+       +#+        */
+/*   By: bedarenn <bedarenn@student.42angouleme.fr> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/06 13:03:13 by bedarenn          #+#    #+#             */
-/*   Updated: 2024/01/21 15:38:24 by bedarenn         ###   ########.fr       */
+/*   Updated: 2024/06/26 15:44:39 by bedarenn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,14 +17,14 @@ int	wati_abs(int nb)
 	return (nb);
 }
 
-int	wati_abs_f(float nb)
+float	wati_abs_f(float nb)
 {
 	if (nb < 0)
 		return (-nb);
 	return (nb);
 }
 
-int	wati_abs_d(double nb)
+double	wati_abs_d(double nb)
 {
 	if (nb < 0)
 		return (-nb);
